eco-sentinals/server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
opencv-python==4.9.0.80
pydantic==2.7.1
scikit-learn==1.4.2
firebase-admin==6.5.0
tensorflow==2.16.1
pytest==8.2.0
httpx==0.27.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
